(MAINT) Initial implementation
This first pass at implementation creates both list and single view
layouts for markdown files in the content/toroidal folder of a hugo
site. The list view will display all of the members of the webring
while the single view generates the navigation menus for member
sites to include as iframes in their own sites.

This implementation includes zero theming or ease of use helpers;
it merely defines a structure for turning member markdown pages into
navigation menus for inclusion in member sites and provides a webring
member site list for convenience.

This implementation includes a first pass at documentation.

Future work will be needed to improve the adoption and usage UX
for the module in order for non-programmers to use it effectively.