(GH-7) Add option to enable link to random member
Prior to this commit, the nav for a toroidal iframe only allowed users
to navigate to the previous/next member in the webring or to the list
of all members on the toroidal host.

This commit adds a new configuration option, **RandomMemberLink** and
modifies the implementation of the single template (which is what the
members will get from their iframe on their own site) to select a
random member from the list of valid members of the webring on click.

Each click is wholly random, so a user can ctrl click the link 2-3
times to load different sites.

Resolves #7.