(MAINT) Overhaul UX for embedding member nav
Prior to this commit, the only way to embed the webring navigation on a
member site was to use the iframe listed in the readme and modify it as
needed; this UX was clunky, fragile, and did not meaningfully help users
who are not already web developers to use the tooling.

This commit substantially overhauls the UX by:

1. Introducing a new Admin page concept and "API" (a json file)
2. Substantially updating the code for the member pages to be more
   robust, stylable, and usable.

Users can now create a minimal admin page as `content/toroidal/admin.md`
and give it the `toroidal/admin` type and set the outputs to `html` and
`toroidal` in the frontmatter to turn display a member administration
page which:

- Includes helpful info about itself at the top
- Includes an entry for each member of the webring with:
  - links to their homepage and the link to the source file for their
    member entry in the webring
  - Snippets of code for embedding as a literal (copy the actual page
    code to their site) or an iframe (dynamically load the navigation
    from the host) with comments and syntax highlighting as well as
    buttons to copy the snippet with and without the comments to make
    copying the snippet to their own site easier
- Makes a minimal JSON "api" available, serving a static file with the
  up-to-date information about the webring's members and the webring
  itself. This is used by the literal embed code, which has scripted
  logic to retrieve this data from the webring host to update any values
  as needed; this enables a member to copy the literal embed once and
  have it always up-to-date without needing to regularly re-retrieve the
  code from the host and update their own site source code.
- Is minimally styled to ensure that it displays relatively neatly.

In addition, any member page can now declare the `toroidal` output type
in its frontmatter to make the json data for that member available.

For the JSON, it is available at `/toroidal/admin/toroidal.json` for the
admin page and `/toroidal/<member-name>/toroidal.json` for any member
pages which opt-in to the behavior.

This commit also substantially reorganizes the partials, decomposing
them to use some `utils` partials to retrieve data and other snippets,
as well as to move the logic for pages into their own partials to keep
the top-level partials as minimal as possible, especially since most of
their code is shared. This should make maintenance a little easier in
the long run.