(GH-4) Support multiple webrings
Prior to this commit, the module could only host a single webring and
was hard-coded to only respect webrings defined in the
`content/toroidal` folder.

This commit enhances the implementation to enable site maintainers to
host a toroidal webring in any folder on their site, including in
subfolders of `content/toroidal`, so long as each folder has an
`_index.md` file with `type` set to `toroidal` in the file's
frontmatter.

This also required overhauling the partials:

- `getMemberPages` was renamed to `getMemberPageLinks` to return the
  URLs to the homepages for every member of the current context's
  webring and modified to use `.Parent.RegularPages` instead of only
  searching for the list of pages in the `toroidal` section at the top
  of the site.
- `getMemberPages` was reimplemented from the same definition but now
  returns the _context_ for those pages instead of their home page URLs.
- `getNextHref` and `getPrevHref` were updated to correctly handle when
  there is no previous/next entry in their section, using the first and
  last items in the list of member pages returned by `getMemberPages`
  (excluding themself) respectively.
- `getPageInfo` now uses `getMemberPages` instead of finding the list of
  pages in the `content/toroidal` section and exposes `memberPages` as
  a variable instead of `toroidalPages`.
- `getWebringName` has been updated to handle using `_index.md` to
  define a webring, favoring settings in that file over the site.
- `getWebrings` is unused but provides a way to check a site for any/all
  defined webrings.

The `toroidal/list` layout was updated to use `getWebringName` instead
of reimplementing the functionality in itself.

The `toroidal/admin/single.toroidal.json` layout was updated to retrieve
member pages from the current section instead of hard-coding a retrieval
from the top level toroidal section.

Finally, the readme was updated to document this functionality.